enh(broker): New output unified sql (#112)
* New output unified_sql
* packaging is now made globally, collect is compiled and then each part is packaged.
* centreon-tests has been completed with tests specific to unified_sql.

REFS: MON-11220

+++ ci/scripts/collect-deb-package.sh
@@ -17,13 +17,20 @@ fi
 
 # mkdir -p centreon-collect
 # cp -rv * centreon-collect
-rm -rf centreon-collect/debian
-rm -rf centreon-collect/build
-tar czpvf centreon-collect-$VERSION.tar.gz centreon-collect
+if [ -d centreon-collect/debian ] ; then
+    rm -rf centreon-collect/debian
+fi
+if [ -d centreon-collect/build ] ; then
+    rm -rf centreon-collect/build
+fi
+tar czpf centreon-collect-$VERSION.tar.gz centreon-collect
 cd centreon-collect/
-mv ci/debian .
+cp -rf ci/debian .
 debmake -f "${AUTHOR}" -e "${AUTHOR_EMAIL}" -u "$VERSION" -r "$RELEASE"
 debuild-pbuilder
 cd ../
+if [ -d "$DISTRIB" ] ; then
+  rm -rf "$DISTRIB"
+fi
 mkdir $DISTRIB
 mv *.deb $DISTRIB/
